Fix sqlite host build for OSX
commit fb30e613139b8836fdc8e81e166cf3a76e5fa17f broke the build process for OSX,
as there's no fdatasync
-#undef __APPLE__ to force the use of fsync

this #if !defined(fdatasync) && !defined(__linux__)
will never eval to true regardless of __linux__ as sqlite3.c
is compiled with -Dfdatasync=fdatasync. Which seems redundant b/c
if the function exists !defined(fdatasync) will eval to false and
fdatasync will stay fdatasync & if it's not present
fdatasync will be defiend as fsync.

Change-Id: Ib97bc466947b84700427f33414ce25c8a8e733b8